apps/api-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
neo4j==5.28.2
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
openai==1.59.3

python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
